Collect Dynamic Connection Level Transmissibilty Multiplier
This commit adds a new data member to the data::Connection
structure,

    double data::Connection::compact_mult

This provides a slot in which the simulator can communicate
transmissibility multiplier values resulting from rock compaction.
The goal is to be able to distinguish background transmissibility
factors from the effective transmissibility factors which include
rock compaction effects.