Enable Empty Structural Copies of Deck Items and Deck Keywords
This commit adds a new member function,

    X emptyStructuralCopy() const

to the DeckKeyword and DeckItem classes.  This member function will
form a "structural" copy of "*this", thus preserving aspects such as
the currently active "Dimensions", but removing all data values.
This is in preparation of adding low-pressure table expansion for
the gas PVT tables (i.e., keywords 'PVDG' and 'PVTG').