Revise Branch Discovery Order for MSW Output
This commit places more conditions on the branch discovery order
than the previous work in commit 6d3ee57dd.  In particular, we now
ensure that branches are discovered ("created") in the order of
increasing segment number of the branch outlet segment.
Furthermore, if multiple branches have the same outlet segment, then
we sort those branches on their increasing branch IDs.

To this end, switch from using a std::queue<int> to using a
std::priority_queue<KickOffPoint>, with the KickOffPoint being a
custom structure holding the kick-off segment, the branch outlet
segment, and the branch ID and a custom operator<() to plug into the
priority_queue<> heap mechanism.

This new sort order changes the result of certain unit tests, but
those changes are expected and desired.