Fix invB from density in co2/h2 store and use density/internalEnergy from co2/h2 directly
We assume that the density and the internal energy of the gas phase
is not affected by the water content.

This avoid usage of gasDensity/gasInternalEnergy based on ideal gas assumptions
implemented in simpleHuDuanH2O.hpp.

Add test that compares liquidEnthalpy of SimpleHuDuanH2O and H2O based on IAPWS