Hotfix for Month String 'FEB' Condition Restart Output Problem
This commit is a quick-fix for a very peculiar problem in the SACN
restart file output array.  We expect that a more complete fix will
involve refactoring.  The problem is, that in conditions such as

    MNTH >= FEB

in an ACTIONX condition block, the current master sources will
unexpectedly interpret the leading 'F' in 'FEB' as introducing a
field-level summary quantity (e.g., FOPR, FPR, or FU_BAR).  This, in
turn, will lead to outputting a value of zero to the corresponding
SACN item instead of the numerical month index for February (= 2).
When restarting a case with this erroneous specification the user
will be greeted with an inscrutable diagnostic message, possibly
referring to internal details of std::unordered_map<> (as retrieved
from eclipseMonthNames())

This commit introduces a special case for date comparisons involving
months to ensure we go into the appropriate branch of the right-hand
side processing.  A more complete fix will, as alluded to earlier,
probably involve refactoring this logic a little bit.

While here, also switch to using WindowedMatrix<double> instead of
WindowedArray<double> as the backing type for sACN_.  This removes
some of the explicit index calculations in computing the correct
view into SACN for the current condition.