Do Not Blindly Remove Blanks in ZLACT Strings
Converting the char[8]-valued elements of the ZLACT restart file
array to std::string will trailing trim blank spaces.  In certain
circumstances, this process will end up removing meaningful blanks
and create run-on keyword data strings that can no longer be parsed
in context.

As an example, the ACTIONX block

    WPIMULT
    'PROD1' 1.23456 3* 1 4 /

might be stored in ZLACT as

    ZLACT[0]: "WPIMULT "
    ZLACT[1]: "        "
    ZLACT[2]: "\n      "
    ZLACT[3]: "'PROD1' "
    ZLACT[4]: "1.23456 "
    ZLACT[5]: "3* 1 4 /"

Upon restart, the string conversion would then trim the trailing
blanks and we would get the keyword specification

    WPIMULT
    'PROD1'1.234563* 1 4 /

which the restart parser object is unable to make sense of.  That,
in turn, means that we might get a diagnostic message of the form

    Error: Problem with keyword WPIMULT
    In <memory string> line 1.
    Malformed floating point number '1.234563*'

and the simulation run would terminate early with

    Error: Unrecoverable errors while loading input

This commit restores the trailing blanks by means of the helper
class PaddedOutputString<8>.  It is more expensive than the current
master sources approach, but it also ensures that we're able to
restart simulation runs which would otherwise run into the problem
described above.