Same region multiplier are not overriden but cumulated.
If source and target region id are the same, the multiplier is
applied to connections within that region and all other regions.
If there are region id pairs with different id and one of those is
the same as for the pair with the same id for both entries, then both
of those multipliers will be applied (and nothing overridden).

This in line with the commercial reference simulator.