Decouple WHISTCTL From History Control Setting on Restart
The current master sources conflates history control with the
WHISTCTL setting, thus failing to respect any control mode setting
in WCONHIST in the case of simulation restart in the historic
period.

This commit decouples any WHISTCTL setting (INTEHEAD[71]) from that
of the control mode setting specified in WCONHIST thereby supporting
both modes when a simulation is restarted in the historic period.
This, in turn, fixes simulation restart on a real field.