Identify Standard Network as Active for NETBALAN Initialisation
This commit extends the NetworkDims::active() function to return
true also for runs using the standard network model (keyword
GRUPNET).  Previously, this member function would return true only
for runs using the extended network model.

This matters, because the initial network balance object's
constructor in member function Schedule::create_first() performs
different actions depending on whether or not the NetworkDims are
"active".  In particular, its pressure convergence limit will be set
to zero unless networks are active in the run.  If there are then no
explicit NETBALAN keywords, then we will try to solve standard
networks to a pressure convergence limit of zero which will
typically fail.

To this end, add two flags to the NetworkDims structure, one to
identify whether or not we're using the extended network model and
one to identify whether or not we're using the standard network
model.  Reimplement the active() member function in terms of these
flags instead of checking the number of network nodes.