Make MINPVV a global array to prevent it from being compressed.
Otherwise any values of cells with ACTNUM pr PORV begin zero would be
removed because cells are not represented.

This was surprising for the test cases that I created, but probably
does not really hurt otherwise. Since we need the global array there
is of course no need to compress in the first place. So we skip it.