Allow simulator to update events of a report step.
This in preparation for correct NEXTSTEP handling during ActionX.
The problem is that there is only a ScheduleState per report
step. This is changed by ACTIONX to have the TUNING_CHANGE event that
we can query from the simulator and react upon. Once we react we want
to prevent future reactios during the same report step. Otherwise
the size specified with NEXTSTEP would be used for every substep.