Internalise RPTSOL Flags for Restart File Output
This commit extends our handling of the RPTSOL keyword when the
'RESTART' flag is present.  In this case, we need to treat remaining
flags as restart file output requests and confer those requests on
to the simulation layer.  We nevertheless need to ensure that those
requests apply only to the time=0 (SEQNUM=0) time point, so
introduce a new helper, solution_only_keywords, to track those flags
which must *not* be forwarded to subsequent time points and remove
those as part of RSTConfig::first().

At this time we use the RPTSCHED handling logic for internalising
RPTSOL mnemonics because the common flags are shared between the two
both for the mnemonics and integer control cases.  We may want/need
to add RPTSOL-specific handling at a later time.

While here, also switch to looping the keywords of the SOLUTION
section, because the ordering of RPTRST and RPTSOL might matter.