Separate Pure From Mixing "Phases" in In-Place Categories
The pure phases are treated differently from the others in the
context of per-cell restart file output (result arrays FIP*, RFIP*,
and SFIP*) so we should have a way excluding those from the general
handling.  To this end, introduce a new helper function,

    Inplace::mixingPhases()

which returns the list of known "phases" apart from the OIL, WATER,
and GAS cases.  We then reimplement the helper function phases() in
terms of mixingPhases().