Add Restart File Output for Fluid-In-Place Arrays
This commit introduces special case logic for outputting per-cell
fluid-in-place arrays.  In particular, we intercept arrays tagged as
'RESTART_SOLUTION' and named as

  * FIP{OIL,WAT,GAS}
  * RFIP{OIL,WAT,GAS}
  * SFIP{OIL,WAT,GAS}

and write those in a separate function (writeFluidInPlace()) instead
of as part of 'writeRegularSolutionVectors()'.  We also add special
compatibility logic to output 'SFIP*' as alias/copy of 'FIP*' if no
'RFIP*' or 'SFIP*' arrays exist in RestartValue::solution.