Fix DATUMRX Keyword Data Model
A default value for the region set name corresponds to the built-in
FIPNUM region set.  While here, also fix the size of the datum depth
portion of each record.  This keyword allows entering multiple depth
values--one for each region--for each region set, so switch the size
type to 'ALL' to accommodate this format.