Internalise Model's Datum Depth from SOLUTION Section
This commit adds a DatumDepth data member to the SimulationConfig
type and initialises this member from the SOLUTION section.  This,
in turn, publishes the model's explicitly or implicitly defined
datum depths--either global or per-region--to client code which may
then use this to calculate depth-corrected phase pressures such as
the BPPO, BPPG, and BPPW summary vectors.