Tidy Up Certain Implementation Files
Mostly by splitting up long lines and re-sorting include statements.
The one big API change here is that we make the UDQ expression
parser private to the implementation file and switch to returning a
unique_ptr<> directly instead of a UDQASTNode which is then used to
form a shared_ptr<>.

All of this is in preparation of adding support for using region
level summary vectors in the defining expressions of field-level
UDQs.