Enable Deferred RegionCache Construction
In particular, pull the construction logic out to a new public
helper function,

    RegionCache::buildCache()

and call this from the body of the constructor.  We need to be able
to initialise/form the cache structure after constructing the object
in the case of UDQs and/or ACTIONX conditions referring to region
level summary vectors which are not otherwise requested.