Don't Index Out of Bounds for GRIDHEAD
When the GRIDHEAD source is a properly constructed restart file
it is okay to assume that GRIDHEAD has at least 25 elements, but in
the context of a manually created regression/unit test that
assumption might not hold.

Play it safe by checking the number of elements before indexing into
the array here.