Enable SummaryConfig Inclusion of UDQ/ACTIONX Vectors
This commit adds a new helper function,

    SummaryConfig::registerRequisiteUDQorActionSummaryKeys

which allows the caller to pass a list (vector<string>) of summary
vector names that are used in the defining expressions of UDQs and
in the condition blocks of ACTIONX keywords.  The SummaryConfig
object will expand the list to a full sequence of summary nodes and
include those in the internal list of "known" vectors.  The function
returns those nodes which did not already exist from processing the
SUMMARY section at construction time.

This is in preparation of generalising the UDQ/ACTIONX support in
Summary.cpp to also include region-level summary vectors.