Expand Set of Categories for UDQ and ACTIONX Summary Vectors
This commit switches to using the new member function

    SummaryConfig::registerRequisiteUDQorActionSummaryKeys()

as a means of constructing the expanded set of summary nodes
referenced by defining expressions for UDQs and in condition blocks
of the ACTIONX keyword.  We furthermore switch to using the
Evaluator::Factory to construct evaluation function objects for
those additional summary nodes.  In principle, this change enables
using all known summary vector categories as part of the defining
expressions in a UDQ, but additional testing is needed before claim
to fully support such usage.

Since this change enables using region-level keywords here, we defer
construction of the 'regionCache_' until we've fully expanded the
set of UDQs and ACTIONX keywords.

The downside to this change is that the registerRequisite* member
function is non-const and therefore requires a mutable SummaryConfig
object.  We have one such object within the EclipseIO container so
this commit switches to taking that object by reference-to-mutable
instead, and move this parameter first in the Summary constructor
argument list.  Update all callers accordingly.