Provide Read-Only Access to Internal SummaryConfig Object
The SummaryConfig object internal to the EclipseIO container might
have a broader set of summary vectors than the one created as part
of processing the SUMMARY section at construction time.  Those
additional vectors would be the ones referenced by UDQs and/or in
the condition blocks of an ACTIONX keyword.  By giving read-only
access to this object, we enable sharing this broader set of vectors
with MPI ranks different from that which holds the EclipseIO object.