Rename UDQSet's Enumerated Item Type
When we add support for using region-level summary vectors in the
defining expressions of field-level UDQs, we need to represent other
collections of (name,number) pairs.  In other words, we don't just
need enumerated well items anymore.

This generalisation means that the existing segment-level UDQSet
support becomes a little more opaque since we now assign the item's
"name" instead of the item's "well", but this is nevertheless
preferable to introducing another type for a very similar concept.