Prepare for Including Region Level Quantities in Field Level UDQs
This commit prepares the UDQContext for including additional summary
vector categories when computing UDQ values.  We introduce a new
wrapper type,

    struct UDQContext::MatcherFactories

whose members are callback functions to create appropriate matcher
objects for the corresponding category.  At this time we support the
original matcher for segment-level summary vectors and the recently
introduced RegionSetMatcher for region-level summary vectors.

We furthermore add a new region set matcher factory parameter to
UDQConfig::eval() in preparation of activating the region-level
support.  The region-level factory is unused at this time, but will
be activated in follow-up work.