Enable Using Region Level Quantities in Field Level UDQs
This commit activates support for using region-level summary vectors
in the defining expressions of field-level UDQs.  In particular, we
add support for evaluating region-level expressions in the UDQ AST
node, retrieving the values of region-level variables as UDQ sets,
and matching region-level expressions in the UDQ context.

Collectively, these changes enable evaluating UDQs defined as

    UDQ
    DEFINE FURE2 (ROIP_RE2 2)+(ROIP_RE2 4)+(ROIP_RE2 5) /
    DEFINE FUGIRD ((0.3*(RGPR_RE2 1)+(RGPR_RE2 2)+(RGPR_RE2 3))/(GEFF TEST)*(1+FUGASX))   /
    DEFINE FUGIRDX (ROIP_RE2 2)/(FURE2)*(0.3*(RGPR_RE2 1)+(RGPR_RE2 2)+(RGPR_RE2 3))/(GEFF TEST)*(1+FUGASX)  /
    /

which occur in the 'UDQ_REG-01' example model in the OPM-Tests
repository.