Relax Assumption That All Connections are Active
This commit allows for the possibility that not all well connections
might be in active cells for the purpose of calculating the WBPn
summary vectors.  This situation might arise when "re-parsing"
portions of the SCHEDULE section following the successful triggering
of an ACTIONX block.  In this case, the well's connections might not
be filtered down to the active set only through a call to member
function Schedule::filterConnections() so it is better to add some
logic here that will filter out the inactive connections.

To this end, add three new members to the PAvgCalculator class:

  - Data member numInputConns_ records the number of original well
    connections--i.e., Well::getConnections().size().  The vector
    of connection-based source locations must be sized according to
    this number of connections.

  - Data member inputConn_ which records a mapping of active->input
    connection IDs.  This is needed to properly identify the correct
    input source location of an active connection

  - Member function pruneInactiveConnections() which filters the
    recorded list of well connections (connections_) down to the set
    of active connections only

We call the latter from the existing member function

    PAvgCalculator::pruneInactiveWBPCells()

since that function knows which cells are active and which are not.