Repair open_well, shut_well, and close_well function.
Added a pointer to a SimulatorUpdate object to the Schedule:
The calls in a PYACTION that trigger a simulator update append this update to the SimulatorUpdate object
of the Schedule. At the end of a PYACTION, this is returned to the Simulator.
Added internal ACTIONX functions for the three functions open_well, shut_well and close_well.
Add a test for open_well and shut_well.