Selectively Bypass FIELD Exclusion From Group Level Keywords
When handling UDQs and, especially, ACTIONX condition blocks, it is
permissible to treat 'FIELD' as a regular group.  In particular, an
ACTIONX condition block might use conditions such as

    GGOR 'FIELD' > 123.4 AND /
    GOPR 'FIELD' < 654.3 /

and we need to be prepared to handle those.  We therefore bypass the
check for group == "FIELD" in keywordG() in this particular context.

We nevertheless preserve the 'FIELD' group exclusion when processing
ordinary group level keywords which don't name specific groups in
the SUMMARY section such as

    GOPR
    /

Future work may elect to include 'FIELD' in this context too.