[bugfix] Fix multiplier handling in the EDIT section.
This is for MULTX, MULTX-, MULTY, MULTY-, MULTZ, MULTZ-.

Previously, EQUALS MULT* would have reset the multiplier of the GRID
section and multiple occurrences of MULT* in the EDIT section would
have been applied cummulatively.

Now the EDIT section has its own multiplier (stored in FieldProperties
with a prefix) which is applied to the one of the GRID
section (without prefix) when we leave the EDIT section.

Hence for multiple occurrences the last one wins and EQUALS MULT* can
be used to only reset some values of the multiplier that will be
applied to the one of the GRID section.