Size Defaulted MULT* Arrays From Active Cells in Output
The EclipseState's input grid's active cells might not coincide with
the set of active cells after grid processing--e.g., due to MINPV.
This commit switches the INIT file's MULT* array output to using the
EclipseIO's notion of the run's active cells and sizes the defaulted
MULT* arrays according to that set.

To this end, move the 'write all mult* minus' flag to the IOConfig
object and form the requisite MULT* array data::Solution object at
the time of output.