Use Unit-Aware Zero Value for Temperature
The bare value 0.0 will be treated as 0 K and output as -273.15 or
-459.67 depending on the unit convention.  This is typically not
what we want.  This commit switches to using an appropriately
dimensioned zero value generated by UnitSystem::to_si(0.0) as the
fallback value for non-existent results.  That value will, in turn,
be output as the value 0.0 in the summary file.  This is more in
line with typical engineering expectations.