updating how to check zero RESV constraint for historic producers
currently, the RESV constraint for a history matching producers will be calculated based on the reservoir
condition at the moment it is used, and we do not have a fixed resv_rate to check against zero. as a result,
we check whether the historic rates for oil, gas and water are all zero to determine whether it is under zero
RESV constraint for history matching producers.