Eliminate Duplicate Map Lookups in SummaryState
A number of get*var() functions were of the form

    if (! has_var(map, var, entity, subEntity)) {
        throw SomeException { ... };
    }

    return map.at(var).at(entity).at(subEntity);

which, if has_var() returns true, ends up looking up the 'var', the
'entity', and the 'subEntity' twice inside 'map'.  This commit
switches to using the appropriate .find() expressions directly.
While there is a considerable cost in code complexity compared to
the original implementation, this is nevertheless justified by not
having to look up each key twice.  Moreover, by using 'find()' we
have more context by which to phrase the diagnostic messages instead
of the generic message that gets generated inside map<>::at().

While here, also switch to using fmt::format() as a means of
constructing composite keys for subscripting the 'values' container.
Finally, try to have single sources of truth for which elements we
reference when updating the various '*values' members and construct
new values in place within the optionals instead of constructing
objects that are subsequently assigned to the optionals.