Make SummaryState Objects Aware of Undefined UDQ Value
This commit switches to calling the SummaryState constructor which
is aware of the value of undefined UDQs directly.  In most cases we
pick the actual value from UDQPARAM(3), with a fallback provided by
the Runspec class if UDQPARAM is not specified, but certain unit
tests which do not care about UDQs just get the value 0.0 as the
second parameter.

This is in preparation of making the undefined UDQ value mandatory
in follow-up work.