Fall Back to Undefined Value for Missing UDQs
This commit switches to using the "undefined" value
(SummaryState::udq_undefined) as a fallback default for undefined or
missing user-defined quantities in the SummaryState::has*() and
SummaryState::get*var() member functions.  In particular, we use
this fallback value if the request is a well UDQ in member functions
(has|get)_well_var(), a group UDQ in (has|get)_group_var(), or a
segment UDQ in (has|get)_segment_var().

Other UDQ categories, e.g., region, will be provided as needed.