Support Large Saturation Function Tables in INIT File
This commit relaxes the INIT file writing code's assumption that
tabulated saturation functions have at most NSSFUN (TABDIMS(3),
default value 20) rows per saturation region.  The rest of the
simulator code does not assume this restriction and there's no
particular reason why the TABDIMS/TAB array should impose this
either.  Moreover, we already relax similar assumptions for the PVT
tables.  Following this commit NSSFUN becomes a "minimum maximum"
number of rows in the TAB array output of the run's saturation
functions.

To this end, switch to returning a three-element tuple<> of the form

    tuple<int, int, vector<double>>

from each function which normalises individual saturation function
keywords.  We then treat this return type as follows

  - Element at index 0 (get<0>(t)) is the number of allocated rows
    in each individual table.  This is usually NSSFUN, but we allow
    input tables larger than this value.  In that case, the number
    of allocated rows will correspond to the maximum number of
    active table rows in the keyword's input data across all
    saturation regions.

  - Element at index 1 (get<1>(t)) is the number of individual
    tables--i.e., the number of saturation regions.  This is
    mostly redundant information but provided as a simplification
    for client code.

  - Element at index 2 (get<2>(t)) is normalised and linearised
    table data for the corresponding saturation functions.

We additionally create a number of intermediate helper functions
named normalise*() which inspect the run's saturation function
family and active phases and calls the appropriate keyword functions
depending on which normalised table we're forming.  This simplifies
the implementation of Tables::addSatFunc().