Collect Well Level Control Limits for Summary Output
This commit adds a transport data structure for collecting well
level control limits, e.g., the oil rate limit of keyword WCONPROD,
for all active wells.  This structure, when populated at the
simulator level, will be the backend for the well level "target"
summary vectors W[OGLVW][IP]RT which are useful for quality control
of running simulation models, as well as giving actual insight into
the active targets which may be affected by keywords such as WTMULT
or user-defined arguments in the control keyword limits.