Trigger Target Update From WTMULT in ACTIONX
We need to inform the simulator about possible changes to wells when
running WTMULT in an ACTIONX block.  Recording the keyword's
affected wells in the handler context enables the simulator to run
any requisite post-action operations such as updating the active
wells' operating targets.  Without the call to

    HandlerContext::affected_well

those post-action operations might not run.