Stop Processing if COPYing From Undefined Source Array
This commit adds defined value checking to the source array from
which we're copying in a COPY operation--either with the COPY or
with the COPYREG keywords.  We count the number of source array
values which do not have the status of 'deck_value' and if this
count is positive, then we throw an OpmInputError exception which
halts further input processing.

As an example of this behaviour, the (deliberately faulty) setup

    EQUALS
    -- SGL not defined prior to this point
     'SGL' 0.0 2* 2* 1 5 /
    /

    COPY
      SGL SWL 2* 2* 4 7 /
    /

will generate an "Error" of the form

    Error: Problem with keyword COPY
    In USER_ERR_COPY_BOX_PARTIALLY_UNDEF.DATA line 195
    Copying from source array SGL in BOX (1-13, 1-22, 4-7)
    would generate an undefined result in 572 blocks of target array SWL.

and stop input processing.