Add macro throw exceptions indicating only problems.
OPM_THROW will increase the error count reported in the PRT file.
For most of the NumericalProblems thrown this is not what we want as
the simulation carries on and errors should halt the simulation. This
commit adds another macro OPM_THROW_PROBLEM that will not increase the
error count but the problem count.

Note that the counters reported in the PRT are often used to find out
whether a simulation succeeded. Hence the error count should only be
nonzero if the simulation did abort.