Properly support scotch as a metis replacement.
If the METIS header is from scotch, then we also need to search for
PtScotch and link to it. For this we now detect PtScotch and adapt the
METIS::METIS target to inlude it. We also force our build system
to use the target instead of the library for linking to also get the
necessary flags automatically.

If PtSCoth is used as a metis replacement then we explicitely request
METIS version 5 APIby defining SCOTCH_METIS_VERSION appropriately.