Don't Needlessly Copy EPS Property Arrays
These copies may have been needed in the past, but that is no longer
the case.  EPS property arrays such as SWL, SOGCR, and SGU are
backed by long-term storage in the FieldPropsManager so we can use
pointers to those instead of storing separate copies of these within
the EclEpsGridProperties structure.

While here, move the EclEpsScalingPointsInfo::update() function into
the .cpp file as it does not need direct access to any data members
of EclEpsScalingPointsInfo.  Finally, move the Leverett J-function
initialisation to a separate member function to reduce the cognitive
load on reading EclEpsScalingPointsInfo<>::extractScaled.