Don't Reference Object Until Index is Known Good
Our restriction on UDAs at FIELD level in GCON* means that we must
not subscript the IUAD vector until we know that 'cnt_iuad' is in
range.  This is the minimal change needed to allow running models
with requests such as

    GCONINJE
    FIELD WATER RATE FURBALL /
    /

but we're still missing the requisite logic that will enable
reliably restarting such models.

The long-term solution to this problem is to add that logic.