Don't Collect UDQ Values if None Exist
This commit switches the AggregateUDQData::dUDx_ members to being
std::optional<> instead of direct WindowedArray<> objects.  This
enables using std::nullopt to represent "no such UDQ category
exists" and makes the output logic in RestartIO.cpp:writeUDQ()
slightly less magic.

While here, also reorder the AggregateUDQData.cpp implementation
file so that the dUdx namespaces appear in alphabetical order.

Finally, move the ig_phase() helper function to the .cpp file since
it does not need to be public and switch the AggregateUDQData
constructor to taking a UDQDims object directly instead of
UDQDims.data().  This enables removing the mostly trivial
createUDQDims() helper function, albeit at a (temporary) cost of
reduced readability in expressions that expected the vector
semantics of its return value.

This work is in preparation of adding support for segment level UDQs
in the restart file.