Extract UDQ Dimensions on Demand
This commmit switches to keeping a reference to the INTEHEAD array
instead of copying the values into internal storage.  This is in
preparation of adding support for segment level UDQs.  Calling the
data() member function will create a vector that's compatible with
the existing set of UDQs, but we do not intend to extend this
collection as other UDQ categories are introduced.