Count All UDQ Variable Types for Output Purposes
This commit ensures that all known UDQ variable types (e.g, field,
group, well, segment) are accounted for in the INTEHEAD restart
vector.  We switch to having an appropriately sized std::array as a
member in InteHEAD::UdqParam represent the counts instead of having
individual members for each variable type.  This is more extensible
as well if we add more categories in the future.  We generate the
category counts to class UDQConfig, since this class already knows
the category counts internally (from UDQConfig::add_node()).  To
this end, add a new member function 'UDQConfig::exportTypeCount()'.

While here, make slight whitespace adjustments to the IUAP and IUAD
counting routines to simplify future maintenance.