Don't Look Up Map Elements Twice
We need the actual element if it exists, so don't go through the
count() member function just to check for existence only to call the
at() member function directly thereafter.  This commit is a
continuation of c720f3296 (PR #4100), but for different map objects.

While here, also reduce the scope of a few iterator objects.