Allow region operator on global 3D fields (poor man's version)
I think for OPERATER there was not even an exception in this case.

We now apply the operations to local storage. After that we prolongate
the results to global storage. Thus for each active cell the value in
the local and global storage match, but no values for inactive cell
will be changed.

We check if the region contains any inactive cells. If this is the case we
issue a warning.