Add Restart File Support for Segment Level UDQs
This commit adds support for outputting and reloading the DUDS
restart file array which holds values of segment level UDQs (SU*) in
an (#segments)-by-(# MS wells)-by-(#segment level UDQs) dense
structure.  Undefined SU* values represented by the usual default
UDQ restart value (-3.0e+20).

As a help to efficiency, we defer to class UDQState to export the
UDQ values pertaining to all segments of a single MS well.  This is
easier than providing an interface based on values for a single
segment at a time.  On the reloading side, we introduce a new member
function UDQVectors::currentSegmentUDQValue() which exposes a
similar range of values for all segments of a single MS well.  While
here, also rework the internals of UDQVectors to resemble the
structure of RstState's UDQ support.