Preserve Value Category When Copying Well Objects
Map_member::get() returns a 'const T&' and it's essential to convey
this information onto the 'back_inserter' lest that helper select
the wrong overload of 'vector<>::push_back()'.  The current code
ended up in the r-value overload vector<>::push_back(T&&) and this
is undesirable when the Well objects must continue to live within
the ScheduleState.