fixed: serialization handling of shared pointers
commit 067d6d44460548e10610b5c753bd2db76bb1887f aimed to
keep the shared nature of shared pointers when deserializing.

the problem is that assigning multiple shared ptrs to the same
raw pointer does not enable any sharing, it only leads to multiple
deletions. fix this by properly holding a shared pointer in the map.

convert the adresses that is serialized to keep track of what pointers were
shared on the sender side to integers (uintptr_t) rather than serializing
them as void pointers. this is clearer and in principle more correct as
it is rather unclear how a serializer should handle a void pointer, and
it only worked by "accident" since std::is_pod_v happens to return true
for a void*