Let ErrorGuard::dump return an error string for logging.
This is needed to let errors in the topology show up in the log files,
too.

Currently, those are only printed to cerr and the simulator will end
with the following line that also appears in the log file:
"Error: Unrecoverable errors while loading input:"

Unfortunately, this line (and the log files) will not contain any
usable error message and there no other message about the error in the
log files.

With this commit we let ErrorGuard::dump return a string with all
errors. Those are slightly reformatted to have error message and file
information on one line. This string will be used in opm-simulators to
print something meaningful to the log files.