Decouple Action::State From Action::Result
This commit introduces a separate storage layer for matching
entities--typically wells--in the Action::State object.  This is in
preparation of making Action::Result more general, and that will
make the Result object harder to serialize.  The new
Action::State::MatchSet type has the usual serialization support,
but is expected to be limited to wells, since that's what's needed
for restart file output purposes.

We switch the internals of Action::State to storing MatchSets
instead and Action::State::result() will now return a pointer to a
MatchSet instead of an optional<Action::Result>.  Update callers
accordingly.

While here, switch Action::State::run_count() to use map<>::find()
instead of map<>::at() since the former enables controlling which
exception type gets thrown as well as the diagnostic message.